Title: 길이가 무한대인 수평선 트랙, Time: 101ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240114/길이가 무한대인 수평선 트랙/horizontal-track-with-infinite-length.cpp
@@ -8,7 +8,7 @@ int main() {
 	int N, T;
 	cin >> N >> T;
 	for (int i = 0; i < N; i++) {
-		int x,v;
+		long long x,v;
 		cin >> x >> v;
 		long long target = x + v*T;
 		while(1){
